Consolidate on isInVariant
