+++ static/src/javascripts/projects/commercial/modules/article-body-adverts.js
@@ -10,8 +10,8 @@ import { trackAdRender } from 'commercial/modules/dfp/track-ad-render';
 import { createSlots } from 'commercial/modules/dfp/create-slots';
 import { commercialFeatures } from 'common/modules/commercial/commercial-features';
 import { initCarrot } from 'commercial/modules/carrot-traffic-driver';
-import { getVariant, isInVariant } from 'common/modules/experiments/ab-core';
 import { commercialAdMobileWebIncrease } from 'common/modules/experiments/tests/commercial-ad-mobile-web-increase.js';
+import { isInVariant } from 'common/modules/experiments/ab-tests';
 
 type AdSize = {
     width: number,
@@ -23,12 +23,8 @@ type Sizes = { desktop: Array<AdSize> };
 
 const isPaidContent = config.get('page.isPaidContent', false);
 
-const isInAdMobileWebVariant = (): boolean => {
-    const variant = getVariant(commercialAdMobileWebIncrease, 'variant');
-    return variant
-        ? isInVariant(commercialAdMobileWebIncrease, variant)
-        : false;
-};
+const isInAdMobileWebVariant = (): boolean =>
+    isInVariant(commercialAdMobileWebIncrease, 'variant');
 
 const adSlotClassSelectorSizes = {
     minAbove: 500,
