+++ static/src/javascripts/projects/commercial/modules/prebid/bid-config.js
@@ -9,7 +9,6 @@ import {
     buildPageTargeting,
 } from 'common/modules/commercial/build-page-targeting';
 import { commercialPrebidSafeframe } from 'common/modules/experiments/tests/commercial-prebid-safeframe';
-import { getVariant, isInVariant } from 'common/modules/experiments/ab-core';
 import type {
     PrebidAdYouLikeParams,
     PrebidAppNexusParams,
@@ -47,11 +46,10 @@ import {
     stripTrailingNumbersAbove1,
 } from './utils';
 import { getAppNexusDirectBidParams, getAppNexusPlacementId } from './appnexus';
+import { isInVariant } from 'common/modules/experiments/ab-tests';
 
-const isInSafeframeTestVariant = (): boolean => {
-    const variant = getVariant(commercialPrebidSafeframe, 'variant');
-    return variant ? isInVariant(commercialPrebidSafeframe, variant) : false;
-};
+const isInSafeframeTestVariant = (): boolean =>
+    isInVariant(commercialPrebidSafeframe, 'variant');
 
 const isDesktopAndArticle =
     getBreakpointKey() === 'D' && config.get('page.contentType') === 'Article';
